feat: Add configuration deliverable templates
This commit adds two new configuration deliverable templates:
`config-deliverable-template.yaml` and
`config-writer-deliverable-template.yaml`. These templates enable the
creation of deliverables for deploying configurations in the cluster.
The deliverables can be used to manage and promote code changes across
different environments using GitOps. The templates provide flexibility
by allowing the use of a monorepo approach or specifying individual
repositories.

The `config-deliverable-template.yaml` template generates a
`Deliverable` resource that includes the necessary configurations and
source code references. It supports GitOps workflows and allows the
configuration of a custom branch, repository, SSH secret, and sub-path.

The `config-writer-deliverable-template.yaml` template generates a
`Runnable` resource that triggers a Tekton TaskRun for writing the
deliverable configurations to the specified repository. It supports both
monorepo and specific repository approaches, and allows the
configuration of the service account, repository, branch, commit
message, and commit files.

These templates provide a standardized and automated mechanism for
managing and deploying configurations, ensuring consistency and reducing
manual effort.